+++ tier1/snappy.cpp
@@ -1025,7 +1025,7 @@ class SnappyIOVecWriter {
       Assert(curr_iov_written_ <= output_iov_[curr_iov_index_].iov_len);
       if (curr_iov_written_ >= output_iov_[curr_iov_index_].iov_len) {
         // This iovec is full. Go to the next one.
-        if (curr_iov_index_ + 1 >= output_iov_count_) {
+        if (curr_iov_index_ + 1 >= (int)output_iov_count_) {
           return false;
         }
         curr_iov_written_ = 0;
@@ -1107,7 +1107,7 @@ class SnappyIOVecWriter {
                                   len);
         if (to_copy == 0) {
           // This iovec is full. Go to the next one.
-          if (curr_iov_index_ + 1 >= output_iov_count_) {
+          if (curr_iov_index_ + 1 >= (int)output_iov_count_) {
             return false;
           }
           ++curr_iov_index_;
